Refactor the chain object tests some
