+++ tests/core/chain-object/test_chain.py
@@ -6,23 +6,27 @@ from eth_utils import decode_hex
 from evm import constants
 from evm.chains.mainnet import MAINNET_GENESIS_HEADER
 from evm.chains.ropsten import ROPSTEN_GENESIS_HEADER
-from evm.estimators.gas import binary_gas_search_1000_tolerance
 from evm.exceptions import (
     TransactionNotFound,
 )
 from evm.vm.forks.frontier.blocks import FrontierBlock
 
-from tests.core.fixtures import (  # noqa: F401
+from tests.core.fixtures import (
     valid_block_rlp,
-    chaindb,
 )
 from tests.core.helpers import (
-    fill_block,
     new_transaction,
 )
 
 
-ADDRESS_2 = b'\0' * 19 + b'\x02'
+@pytest.fixture
+def chain(chain_without_block_validation):
+    return chain_without_block_validation
+
+
+@pytest.fixture
+def valid_chain(chain_with_block_validation):
+    return chain_with_block_validation
 
 
 @pytest.fixture()
@@ -80,133 +84,6 @@ def test_empty_transaction_lookups(chain):
 
     with pytest.raises(TransactionNotFound):
         chain.get_pending_transaction(b'\0' * 32)
-
-
-@pytest.fixture(
-    params=[True, False])
-def unsigned_or_signed_tx(request):
-    def tx(
-            vm,
-            from_,
-            to,
-            amount,
-            private_key=None,
-            gas_price=10,
-            gas=100000,
-            data=b''):
-
-        if request.param:
-            return new_transaction(
-                vm,
-                from_,
-                to,
-                amount,
-                private_key,
-                gas_price,
-                gas,
-                data)
-
-        else:
-            return new_transaction(
-                vm,
-                from_,
-                to,
-                amount,
-                private_key=None,
-                gas_price=gas_price,
-                gas=gas,
-                data=data)
-    return tx
-
-
-@pytest.mark.parametrize(
-    'data, gas_estimator, to, on_pending, expected',
-    (
-        (b'', None, None, True, 21000),
-        (b'', None, None, False, 21000),
-        (b'\xff' * 10, None, None, True, 21680),
-        (b'\xff' * 10, None, None, False, 21680),
-        # sha3 precompile
-        (b'\xff' * 32, None, ADDRESS_2, True, 35381),
-        (b'\xff' * 32, None, ADDRESS_2, False, 35369),
-        (b'\xff' * 320, None, ADDRESS_2, True, 54888),
-        # 1000_tolerance binary search
-        (b'\xff' * 32, binary_gas_search_1000_tolerance, ADDRESS_2, True, 23938),
-    ),
-    ids=[
-        'simple default pending',
-        'simple default',
-        '10 bytes default pending',
-        '10 bytes default',
-        'sha3 precompile 32 bytes default pending',
-        'sha3 precompile 32 bytes default',
-        'sha3 precompile 320 bytes default pending',
-        'sha3 precompile 32 bytes 1000_tolerance binary pending',
-    ],
-)
-def test_estimate_gas(
-        chain,
-        data,
-        gas_estimator,
-        to,
-        on_pending,
-        expected,
-        funded_address,
-        funded_address_private_key,
-        unsigned_or_signed_tx):
-    if gas_estimator:
-        chain.gas_estimator = gas_estimator
-    vm = chain.get_vm()
-    if to:
-        recipient = to
-    else:
-        recipient = decode_hex('0xa94f5374fce5edbc8e2a8697c15331677e6ebf0c')
-    amount = 100
-    from_ = funded_address
-    tx = unsigned_or_signed_tx(vm, from_, recipient, amount, funded_address_private_key, data=data)
-    if on_pending:
-        # estimate on *pending* block
-        assert chain.estimate_gas(tx, chain.header) == expected
-    else:
-        # estimates on top of *latest* block
-        assert chain.estimate_gas(tx) == expected
-        # these are long, so now that we know the exact numbers let's skip the repeat test
-        # assert chain.estimate_gas(tx, chain.get_canonical_head()) == expected
-
-
-def test_estimate_gas_on_full_block(chain, funded_address_private_key, funded_address):
-
-    def estimation_txn(chain, from_, from_key, data):
-        to = decode_hex('0xa94f5374fce5edbc8e2a8697c15331677e6ebf0c')
-        gas = chain.header.gas_limit
-        amount = 200
-        vm = chain.get_vm()
-        return new_transaction(vm, from_, to, amount, from_key, gas=gas, data=data)
-
-    from_ = funded_address
-    from_key = funded_address_private_key
-    garbage_data = b"""
-        fill up the block much faster because this transaction contains a bunch of extra
-        garbage_data, which doesn't add to execution time, just the gas costs
-    """ * 30
-    gas = 375000
-
-    # fill the canonical head
-    fill_block(chain, from_, from_key, gas, garbage_data)
-    chain.import_block(chain.get_vm().block)
-
-    # build a transaction to estimate gas for
-    next_canonical_tx = estimation_txn(chain, from_, from_key, data=garbage_data * 2)
-
-    assert chain.estimate_gas(next_canonical_tx) == 722760
-
-    # fill the pending block
-    fill_block(chain, from_, from_key, gas, garbage_data)
-
-    # build a transaction to estimate gas for
-    next_pending_tx = estimation_txn(chain, from_, from_key, data=garbage_data * 2)
-
-    assert chain.estimate_gas(next_pending_tx, chain.header) == 722760
 
 
 def test_canonical_chain(valid_chain):
