--- tests/core/vm/conftest.py
@@ -1,10 +1,0 @@
-from tests.core.fixtures import (  # noqa: F401
-    # Constant
-    funded_address,
-    funded_address_private_key,
-    funded_address_initial_balance,
-    # Chain
-    chaindb,
-    chain as valid_chain,
-    chain_without_block_validation as chain,
-)
